(fix): create `types.py` file for `ExtensionNamespace`
